requirements.txt
streamlit
pymongo
scikit-learn
pandas
flask
python-dotenv
